Move query count to a test utilitary file
